+++ test/validate.spec.php
@@ -7,4 +7,7 @@ describe('Krak Validation', function() {
     describe('Violation', function() {
         require_once __DIR__ . '/violation.php';
     });
+    describe('ViolationCollection', function() {
+        require_once __DIR__ . '/violation-collection.php';
+    });
 });